server/requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai>=1.30.0
spacy==3.7.2
python-dotenv==1.0.0
scikit-learn==1.5.0
beautifulsoup4==4.12.3
requests==2.32.5
numpy==1.26.4
